Data.Functor.Exp

Description

Documentation

class Exp f where

Exponential/invariant functors.

Methods

emap :: (a -> b) -> (b -> a) -> f a -> f b

class ExpProfunctor p where

Exponential/invariant profunctors.

Methods

elmap :: (a -> b) -> (b -> a) -> p a c -> p b c

ermap :: (a -> b) -> (b -> a) -> p c a -> p c b

class Exp f => Combinable f where

The exponential analogue to Applicative covariant functors or Divisible contravariant functors.

Methods

combineId :: a -> f a

combine :: f a -> f b -> f (a, b)

class Combinable f => Pickable f where

The exponential analogue to Alternative and Divisble.

Methods

pickId :: f Void

pick :: f a -> f b -> f (Either a b)

class Divisible f => Contramonad f where

A sort-of monad analogue for contravariant functors. If 'f a' consumes an a, and there is a function from a to a consumer f b, pairbind makes a consumer for an f (a, b).

Methods

pairbind :: f a -> (a -> f b) -> f (a, b)

class Combinable f => Constructable f where

A sort-of- monad analogue for exponential functors, similar to Contramonad.

Methods

construct :: f a -> (a -> f b) -> f (a, b)