Data.Stream

Description

Documentation

class Stream s where

Anything that can be uconsed or appended to, and that has an empty element. You can think of this as a monoid that can have the first element split off.

Associated Types

type Token s

Methods

emptyStream :: s

appendStream :: s -> Token s -> s

unconsStream :: s -> Maybe (Token s, s)