Transceiver.Combinators

Description

Documentation

manyE :: Pickable f => f a -> f [a]

Syntax combinator similar to many, for zero or more as, until the usable IO is exhausted.

[a a .. a]

someE :: Pickable f => f a -> f (NonEmpty a)

Syntax combinator similar to some, for one or more as, until the usable IO is exhausted.

a [a a .. a]

untilE :: Pickable f => f a -> f b -> f ([a], b)

Syntax combinator for zero or more as, until the first b.

[a a .. a] b

until1E :: Pickable f => f a -> f b -> f (NonEmpty a, b)

Syntax combinator for one or more as, until the first b.

a [a a .. a] b

optional :: Pickable f => f a -> f (Maybe a)

Syntax combinator for an optional element. When parsing, will return Nothing when a fails.

[a]

repeatN :: Combinable f => Int -> f a -> f [a]

Syntax for n repeated instances of a.

a1 a2 ... an

constant :: Exp f => a -> f a -> f ()

Disregard the information from the covariant functor, and replace the information in the contravariaint functor with x.

For syntax combinators, this ignores the parsed token and prints the same thing every time.

padding :: Combinable f => Int -> a -> f a -> f ()

Syntax combinator that disregards n ts, and prints n fs.

Parsing:

_1 _2 ... _n

Printing:

f1 f2 ... f3

sequenceL :: Combinable f => f a -> f () -> f a

Assuming a carries no information, sequence a and b, like <*

a b

sequenceR :: Combinable f => f () -> f a -> f a

Assuming b carries no information, sequence a and b, like *>

a b

sepBy :: Pickable f => f a -> f () -> f [a]

sepBy1 :: Pickable f => f a -> f () -> f (NonEmpty a)

between :: Combinable f => f () -> f () -> f a -> f a