Transceiver.Fields

Description

Documentation

fld :: forall (k :: Symbol) v f. Exp f => f v -> f (NMap '[k :-> v])

Isomorphism from a to an NMap with a single, named field. Use -XTypeApplications to set the field name.

>>> :t fld @"hello" token
Syntax s (NMap '["hello" ':-> Token s])

con :: forall k v f. Exp f => f v -> f (NSum '[k :-> v])

Isomorphism from a to an NSum with a single, named constructor. Use -XTypeApplications to set the field name.

>>> :t con @"hello" token
Syntax s (NSum '["hello" ':-> Token s])

(|&|) :: forall (xs :: [Mapping Symbol *]) (ys :: [Mapping Symbol *]) f. (SingI xs, SingI ys, Combinable f) => f (NMap xs) -> f (NMap ys) -> f (NMap (Union xs ys)) infixr 5

combine two exponential functors of NMaps. (The order of the arguments determines the order of the serialized or parsed result, but not the order of the fields in the type.)

(|/|) :: forall (xs :: [Mapping Symbol *]) (ys :: [Mapping Symbol *]) f. (SingI xs, SingI ys, Pickable f) => f (NSum xs) -> f (NSum ys) -> f (NSum (Union xs ys)) infixr 3

pick two exponential functors of NSums. (The order of the arguments determines the order of the parsed branches are tried, but not the order of the fields in the type.)

(=&) :: forall n a f. Exp f => f a -> FieldLabel n -> f (NMap '[n :-> a]) infixl 6

(=/) :: forall n a f. Exp f => f a -> FieldLabel n -> f (NSum '[n :-> a]) infixl 4

pmap :: (Exp f, Generic a, GenProduct (Rep a)) => f (NMap (GProduct (Rep a))) -> f a

smap :: (Exp f, Generic a, GenSum (Rep a)) => f (NSum (GSum (Rep a))) -> f a

class Generic a

Representable types of kind *. This class is derivable in GHC with the DeriveGeneric flag on.

A Generic instance must satisfy the following laws:

from . to ≡ id
to . from ≡ id

Minimal complete definition

from, to