Transceiver.Iso

Description

Documentation

gmap :: (Exp f, Generic a, Generalizable' (Rep a)) => f (Generalized (Rep a)) -> f a

Maps an exponential functor from its generic version to its specialized version.

The generic versions of datatypes are the ones parsed natively by the |*| and |+| combinators. That is, product types like A B C D are represented as (A, (B, (C, D))), and coproduct types like A | B | C are represented as Either A (Either B C).

If your datatype has a Generic instance, and your syntax combinators follow the same order as the constructors, inserting a gmap in front should be all that's needed.