Transceiver.Parser

Description

Documentation

newtype Parser s a

Since not every input string is a valid representation of the output type, parsers must return a Maybe. This is a functor covariant in its argmuent.

Constructors

Parser (s -> Maybe (a, s))

parseToken :: Stream s => Parser s (Token s)

Parse one token from the input stream, failing only if the input is empty.

parseEof :: Stream s => Parser s ()

Succeed if the input stream is empty, fail otherwise.

parseSatisfy :: Stream s => Parser s a -> (a -> Bool) -> Parser s a

Succeed if the result of the parse matches the predicate.