Transceiver.Printer

Description

Documentation

newtype Printer s a

Since it's assumed that all values of the type being printed are valid, it's assumed they all have a valid representation in the output stream.

Constructors

Printer (a -> s -> s)

printToken :: Stream s => Printer s (Token s)

Print a single token by appending it to the output stream.

printEof :: Stream s => Printer s ()

Do nothing at all to the output stream.

printInsert :: Stream s => Token s -> Printer s ()

Insert x into the output stream, without actually needing it as an input.