Transceiver.Syntax

Description

Documentation

type Syntax s a = Product (Parser s) (Printer s) a

The main type representing a syntax combinator: it's just a pair with a parser and a printer. This isn't a newtype so users can make them manually from components with Pair or split them apart with pattern matching.

token :: Stream s => Syntax s (Token s)

Syntax for a single token in the IO stream.

runParser' :: Syntax s a -> s -> Maybe (a, s)

Parse the input stream s, and if it succeeds, return the resulting a and the unconsumed input.

runPrinter' :: Syntax s a -> a -> s -> s

Print an a, appending the results of the print with the given stream.

runParser :: Syntax s a -> s -> Maybe a

Same as runParser, but discard the unused stream.

runPrinter :: Stream s => Syntax s a -> a -> s

Same as runPrinter, but append to the empty stream.

eof :: Stream s => Syntax s ()

Succeeds in parsing only the empty stream, prints nothing.

literal :: (Stream s, Eq (Token s)) => s -> Syntax s ()

Syntax for an exact match of a sequence of tokens, like exactToken.

exactToken :: (Stream s, Eq (Token s)) => Token s -> Syntax s ()

Syntax for an exact match of a single token. Produces/consumes an '()' because an exact match carries no information.

satisfy :: Stream s => (a -> Bool) -> Product (Parser s) g a -> Product (Parser s) g a

Only suceed in parsing if the predicate f is satisfied. This assumes that the predicate will always hold when printing.